bot/pyproject.toml
[tool.poetry]
name = "discord-bot-python-runner"
version = "0.1.0"
description = ""
authors = ["mkizka"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
"discord.py-stubs" = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
